+++ lib/src/components/styled/SendInput.tsx
@@ -23,6 +23,7 @@ import AudioRecorder from '../InputComponents/AudioRecorder';
 import { IConfig } from '../../types/types';
 import Button from './Button';
 import { AttachIcon, FileIcon, RemoveIcon, SendIcon } from '../../assets/icons';
+import { resolveIconColor } from '../../helpers/resolveIconColor';
 import { parseMessageBody } from '../../helpers/parseMessageBody';
 
 export interface SendInputProps {
@@ -271,9 +272,9 @@ const SendInput: React.FC<SendInputProps> = ({
     } else if (fileType === 'video') {
       return <VideoPreview src={fileUrl} controls />;
     } else {
-      return <FileIcon alt={file.name} />;
+      return <FileIcon alt={file.name} color={resolveIconColor(config)} />;
     }
-  }, []);
+  }, [config]);
 
   const memoizedFilePreviews = useMemo(() => {
     return filePreviews.map(
@@ -307,7 +308,7 @@ const SendInput: React.FC<SendInputProps> = ({
               <Button
                 onClick={handleAttachClick}
                 disabled={false}
-                EndIcon={<AttachIcon />}
+                EndIcon={<AttachIcon color={resolveIconColor(config)} />}
               />
             )}
             {multiline ? (